+++ CHANGELOG.md
@@ -14,8 +14,9 @@
 - Added `Widget.options.reportPageId` option
 - Improved `Widget.setPageHash()` method to detect changes based on decoded page id, ref and params and re-apply new values if needed
 - Fixed `table` to not override `data` when column config in `cols` object have no `content` property
-- Fixed `table` to apply `when` on column config to entire column
-- Changed `when` and `whenData` to not apply when `undefined` (like not set at all)
+- Changed `table` to apply `when` on column config to entire column
+- Changed `table` to not render a cell content depending on `whenData` in column config
+- Changed `when` and `whenData` to not apply when `undefined` or not an own property (like not set at all)
 
 ## 1.0.0-beta.31 (23-04-2020)